Fix performance regression
